Merged revisions 63-105,107-125 via svnmerge from  https://scm.sipfoundry.org/rep/sipXconfig/branches/post-2.6.1
NOTE: This was done by hand.  svnmerge -L post-2.6.1 had many conflicts
which may have been caused by bidirectional merge. This commit was the
result of executing this command:

     svn merge \
      https://scm.sipfoundry.org/rep/sipXconfig/main \  
      https://scm.sipfoundry.org/rep/sipXconfig/branches/post-2.6.1 .

And there was no conflicts or addtional changes that had to be made. 
There was a merge from main to 2.6.1 done just prior to this merge from
2.6.1 to main



git-svn-id: http://scm.sipfoundry.org/rep/sipX/old/sipXconfig/main@3150 a612230a-c5fa-0310-af8b-88eea846685b