Fixed two cases where the proxy falsely assumed strict routing. These cases were intensionally put in to fix problems caused by user agents that strict routed but did not preserve the lr parameter in the route header.
The first case was when a message was loose routed and sent to
the proxy even though there was no route for the proxy in the 
message.  This often occurs when user agents force messages to
be sent to its configured outbound proxy event though the
Record-Route does not include the proxy in the route set.

The other case occurs when a user agent puts a pre-set route
in a request that is not part of a dialog.  The URI and the
first route both go to the proxy.  The proxy assumed this to
be a strict routing case as this was what the requests looked
like if they dropped the lr parameter in the route.  We are
now assuming that user agents no longer drop the lr parameter
or if they do, tough luck.

See also: http://track.sipfoundry.org/browse/XPR-9




git-svn-id: http://scm.sipfoundry.org/rep/sipX/old/sipXproxy/main@2248 a612230a-c5fa-0310-af8b-88eea846685b