Temporary fix: #ifdef out the tests:      CallManagerTest::testUATeardown      CallManagerTest::testLineMgrUATeardown      CallManagerTest::testRefreshMgrUATeardown so that sipX will build.
Put parentheses around #define expression.

Do not treat an absent port number as equivalent to the default port number.
http://track.sipfoundry.org/browse/XSL-96

Remove sipXconference from the process definitions for the time being, because
it is not done yet.  If it appears in ProcessDefinitions.xml, then
the config UI will display it as a process (that has not been started).


git-svn-id: http://scm.sipfoundry.org/rep/sipX/main/sipXproxy@4931 a612230a-c5fa-0310-af8b-88eea846685b