Workaround for XSL-49 - ACK misrouting and logging "using UDP"
The incidence of the log message about "using UDP" is reduced by this
to only those cases where we really are just presumptively using UDP;
previously it was sometimes logged even though a different protocol
was being used because the transaction had a record in
lastSentProtocol.

There is still a problem here - this path gets hit in the authproxy
for ACKs to 2xx responses, and they may be routed incorrectly when the
real target doesn't accept UDP.  See the comments added by this fix.


git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@5247 a612230a-c5fa-0310-af8b-88eea846685b