Finish reverting rev. 3920, which was a complete mistake. sipXregistry/include/RegisterHook.h was deleted in rev. 3926.
Add log-debug to the PHONIES list.

Allow TiXmlBase::SkipWhiteSpace to work correctly when *p == '\0' by
returning successfully.

Revert rev 3891 (except for an improvement in a debugging print).
That rev was to work around the bug in XPL-127.  Now that rev 4040 has fixed
that bug, remove the workaround.

Add test for TinyXML problem XPL-127.

For the time being, force the "no early-only" behavior in all call
pick-ups.

Fix handling of parameters on the remote contact address in dialog
event notices.

Add workaround so Snom phones can pick up calls.

Split SipClient::isReadyToRead into two functions:
SipClient::isReadyToRead, which is a non-blocking test for socket-ready-to-read,
and SipClient::waitForReadyToRead, which waits.
http://track.sipfoundry.org/browse/XSL-111


git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@5450 a612230a-c5fa-0310-af8b-88eea846685b