XML-RPC interface for remote OsConfigDb manipulation.
sipXcommserverLib now has a general mechanism to remotely manipulate
an OsConfigDb - see sipXcommserverLib/src/configrpc/README for a
description of the XML-RPC interface it implements.

bbridge now provides remote configuration access based on that
mechanism.  The security of that access is not done yet (some
work is needed in the SSL/HTTPS/XML-RPC interface for this).

Changed the bbridge configuration parameter names to facilitate
modular configuration of conferences and add some new parameters.
Most of them are not used yet, but they are documented now and
initialized.


git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@5586 a612230a-c5fa-0310-af8b-88eea846685b