[XSL-72] Validate peer certs and provide connection security attribute
Adds methods to OsConnectionSocket and OsSSLConnectionSocket that
allow access to the security characteristics of the connection.

Records those attributes in the HttpRequestContext so that they are
available to anything layered on top of HTTP.

Adds checking of the security for the ConfigRPC mechanism for remote
manipulation of an OsConfigDb.

Use the security in the bbridge configuration.

http://track.sipfoundry.org/browse/XSL-72


git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@5640 a612230a-c5fa-0310-af8b-88eea846685b