Prevent ./build from hanging when no labels are specified.
Allow "-p Project+" to mean "Project" and all following projects in
the standard build order.

At the beginning of a run, print the order projects will be built.

Get the tab characters out of ./build.
In some places, I had to use the $'...\t...' construction, which may
not be as portable as we'd like, but it appears to be Posix standard
on the bash manual page.

Modified to comply with the new dist tree


git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@6051 a612230a-c5fa-0310-af8b-88eea846685b