Merge 6730,6732,6734 from sipfoundry 3.4:
Fix XPR-154: In CDR , HA setup, if the distributed machine is down CDR is not logged .
If we get an error connecting to a CSE DB, then catch it, report it, and at least try to do CDRs for the DBs that we are able to reach.

Add a variant of the call resolver config file for use in development

Fix the way the load path is handled to eliminate annoying warnings about constants already having been defined.
Add unit tests for DatabaseUtils.


git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@6735 a612230a-c5fa-0310-af8b-88eea846685b