mergefrom-incremental pingtel-main 4943
Added the check for dialog handle to prevent the crash of
unsubscribing due to the event of not being successfully subscribed at
the first place.

Added some checks for the active agent pointer before using it to
prevent the crash.

Fixed the issue of putting the routing call in the queue (ENG-346) and
also added stopAudio when a call is being dropped.
http://paradise.pingtel.com:8080/browse/ENG-346

Cleaned up all the timers in the destructor.

Fixed the issue of crash during the ACD shutdown.
http://paradise.pingtel.com:8080/browse/ENG-374


git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@7663 a612230a-c5fa-0310-af8b-88eea846685b