mergefrom pingtel-main 5813
Also, revamp SipSubscriptionMgr to create and use its own
SipDialogMgr, rather than accepting one from its creator.  The latter
policy appears to be more efficient, but is hard to make work well, as
the creator doesn't know when SipSubscriptionMgr is done with the
SipDialogMgr, and most of the time has to create a SipDialogMgr just
for SipSubscriptionMgr.
fixed crash in sipXtackLibTest.
SipDialogManager is owned by the Subscription Manager, so, it should
not be deleted.


git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@7681 a612230a-c5fa-0310-af8b-88eea846685b