mergefrom pingtel-main 6936,7277,6085
Move a couple of classes to be subclasses of UtlContainableAtomic
rather than UtlContainable.
Update a bunch of classes to use the class::getContainableType() and
class:TYPE members according to our conventions.

Correct how GRUUs are generated.
Fix how libsipXcommserverTest.la is searched for, so it can be found
even when it is not installed.


git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@7707 a612230a-c5fa-0310-af8b-88eea846685b