mergefrom sipx-main 7723
mergefrom pingtel-main 7063,7092
In CpPeerCall::handleGetSession and handleGetInvite, revise the search
for the call in question to always compare the call-ID and to compare
the address both with its tag and without, because the address is
recorded in the SipConnection without its tag before the call is
established.  The previous test always used the tag, so both methods
failed when applied to ringing calls.
Improve debugging prints.
Also, make some edits to new code in this branch to match the changes
in the CpPeerCall functions.


git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/3.6@7734 a612230a-c5fa-0310-af8b-88eea846685b