mergefrom pingtel-main 7460
mergefrom pingtel-3.6 7441-7448,7452,7457-7458
Extract the message-processing code from SipAaa::handleMessage as a
separate method, SipAaa::proxyMessage, as has been done before in the
forking proxy.
Partial revert of rev. 6997, so that the tags are taken into account
in the route-signature mechanism.
It turns out that the original design allows requests in the "forward"
direction (the direction of the dialog-creating request) to be
authenticated, and requests in the "backward" direction to never be
challenged.
This mechanism doesn't always work, though, and restores the original
problem seen in XPR-64.
Changes to not require authorization of in-dialog requests.
http://track.sipfoundry.org/browse/XPR-197
Add UrlMapping::loadMappingsString to load the mappings from a
supplied string, rather than reading them from a file, for testing
purposes.
Type read-only UtlString arguments as 'const UtlString&'.
Add tests for the straightforward versions of all the cases we care
about.


git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@7848 a612230a-c5fa-0310-af8b-88eea846685b