mergefrom sipx 7763,7847,7854
Fixed call-id generator so that it does not require 64 bit 
printf format: %lld as this seems to have some problems on
some platforms.

XPB-836 Martin's fix for Apache home dir on FC6
not sure if this is the right thing to do and if someone has a better idea feel free to fix it 
I just want to prepare for FC6 build loop

Yet another version of CpCallManager::getNewCallId.
This one removes the need for using long long int's (and the attendant
portability hassles), shortens the Call-IDs slightly, but maintains
the other desirable properties of the previous version.
Add a test to check that successive Call-IDs aren't too similar.


git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/user_dworley_redirectors@7856 a612230a-c5fa-0310-af8b-88eea846685b