mergefrom pingtel-main 7480
Delete unused (and dangling) link.
Fix the test which detects (and does not pass to the upper layers)
DTMF events which indicate only the prolongation of a key that is
already down.
Clean up comments.
Solve the problem with Polycom phones (which have an error if a
re-INVITE arrives while a DTMF key is down) in two ways:
Update ParkedCallObject to call CallManager::transfer_blind with the
flag indicating that the call should not be put on hold before the
REFER is sent.
Perform the transfer-to-parker operation upon receiving the keyup
event, not the keydown event.


git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@7898 a612230a-c5fa-0310-af8b-88eea846685b