Add ability to disable watchdog-process "process.d" XML files from being processed.
Changed the watchdog schema to add an "enable" attribute to the top level "watchdog-process" element.
 
This causes the file to be ignored if not enabled (enable="false").  
Thus it acts as if it doesn't exist in the process.d directory, and so 
doesn't ever get started, and better yet doesn't even show up as a service, 
running, stopped or otherwise.  It has an advantage over just deleting or 
renaming the file in that subsequent updates will not overwrite the file as 
it exists and has been modified.

This simplifies the HA configuration as well, as the file only needs to be
changed in one location, not two.

Fixes http://track.sipfoundry.org/browse/XPB-823
Fixes http://track.sipfoundry.org/browse/XPB-837

--Woof!


git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/3.6@7938 a612230a-c5fa-0310-af8b-88eea846685b