Video calls stopped working after r8182 fix.
It's strange for me, but OsSocket::remoteHostPort is not set to real value if object have OsNatDatagram type. As I see from the code, remoteHostPort is set to value, passed to OsNatDatagramSocket constructor and then never refreshed. So OsSocket::getRemoteHostPort() always return port number, passed to OsNatDatagram constructor, which is 0 in this case.


git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/sipXtapi-media-update@8184 a612230a-c5fa-0310-af8b-88eea846685b