Added mutex locking in RIFF file recording
When it came time to stop a recording, one thread was trying to update 
the RIFF header on the file, while at the same time one thread was still
adding audio data to the end of the file.  

At times this caused mismatches between the actual file size and what was 
reported in the header.  Other times it created a corrupted header
because the audio data was written just as the update header code did
an "lseek()" to set the file pointer back.

Fixes ENG-723 "Some voice mail messages appear to the user as blank audio".


--Woof!


git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@8309 a612230a-c5fa-0310-af8b-88eea846685b