Correct media render stall when EXACTLY 16 seconds of media is queued.
With 1600 100 mS buffers, only 16 seconds can be queued before an event
starting the render needs to be triggered.  However, the "end of frames" 
message is also queued in the same queue, and no check was done when trying 
to send that message.  If all 1600 buffers were used for media, when it came 
time to send the "end of frames" message it blocked, being unable to get
a buffer.  But no event triggered the other side to start reading the queue 
so the render just hung.

Resolved this by sending a throttle event on the "end of frames" messages
as well if needed.  This will start the other side reading the queue which 
will eventually free up a buffer and prevent the stall.

This issue is also known as "The .WAV of Doom" because if a .wav file of
EXTACTLY 16 seconds (256000 bytes + 44 byte RIFF header) is in a voice
mailbox, then the media server will just hang when that message is played.

Fixes XPB-736

--Woof!


git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@9103 a612230a-c5fa-0310-af8b-88eea846685b