Fix "hang" or "hangup" when 0 length .wav files end up in voicemail.
If a 0 length .wav file is encountered, instead of throwing an error
which causes the call to be dropped, play a quiet "click" instead.

Also handle the case where a 0 length or other "problem" .wav file is
forwarded would cause a malformed .wav file that was the result of the
concatenation.  The new .wav file contains as much as possible, and has
the correct header.  (Previously the header would be wrong).

Fixes XPB-923 (trying to listen to a voicemail msg of 0 length causes 
the listener to exit voicemail)

--Woof!


git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@9167 a612230a-c5fa-0310-af8b-88eea846685b